변경 : Player class 이동 및 View 를 위한 display 변수 추가
