+++ src/main/kotlin/blackjack/domain/card/Pattern.kt
@@ -1,5 +1,5 @@
 package blackjack.domain.card
 
-enum class Pattern {
-    CLOVER, DIAMOND, HEART, SPADE,
+enum class Pattern(val display: String) {
+    CLOVER("클로버"), DIAMOND("다이아몬드"), HEART("하트"), SPADE("스페이드"),
 }
